Feat: Added testing folder and Makefile.
