+++ requirements/requirements.in
@@ -1,0 +1,4 @@
+pytest
+mypy
+tox
+ipython
