+++ requirements/requirements.txt
@@ -1,0 +1,44 @@
+#
+# This file is autogenerated by pip-compile
+# To update, run:
+#
+#    pip-compile requirements/requirements.in
+#
+appdirs==1.4.4            # via virtualenv
+appnope==0.1.0            # via ipython
+attrs==20.2.0             # via pytest
+backcall==0.2.0           # via ipython
+decorator==4.4.2          # via ipython, traitlets
+distlib==0.3.1            # via virtualenv
+filelock==3.0.12          # via tox, virtualenv
+importlib-metadata==2.0.0  # via pluggy, pytest, tox, virtualenv
+importlib-resources==3.0.0  # via virtualenv
+iniconfig==1.0.1          # via pytest
+ipython-genutils==0.2.0   # via traitlets
+ipython==7.16.1           # via -r requirements/requirements.in
+jedi==0.17.2              # via ipython
+mypy-extensions==0.4.3    # via mypy
+mypy==0.790               # via -r requirements/requirements.in
+packaging==20.4           # via pytest, tox
+parso==0.7.1              # via jedi
+pexpect==4.8.0            # via ipython
+pickleshare==0.7.5        # via ipython
+pluggy==0.13.1            # via pytest, tox
+prompt-toolkit==3.0.8     # via ipython
+ptyprocess==0.6.0         # via pexpect
+py==1.9.0                 # via pytest, tox
+pygments==2.7.1           # via ipython
+pyparsing==2.4.7          # via packaging
+pytest==6.1.1             # via -r requirements/requirements.in
+six==1.15.0               # via packaging, tox, traitlets, virtualenv
+toml==0.10.1              # via pytest, tox
+tox==3.20.1               # via -r requirements/requirements.in
+traitlets==4.3.3          # via ipython
+typed-ast==1.4.1          # via mypy
+typing-extensions==3.7.4.3  # via mypy
+virtualenv==20.0.34       # via tox
+wcwidth==0.2.5            # via prompt-toolkit
+zipp==3.3.0               # via importlib-metadata, importlib-resources
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
